Rename swircheritem to item and delete confpage
