--- spyder/plugins/switcher/confpage.py
@@ -1,18 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# Copyright © Spyder Project Contributors
-# Licensed under the terms of the MIT License
-# (see spyder/__init__.py for details)
-
-"""Switcher Preferences Page."""
-
-from spyder.api.preferences import PluginConfigPage
-from spyder.api.translations import _
-
-
-class SwitcherConfigPage(PluginConfigPage):
-
-    # --- PluginConfigPage API
-    # ------------------------------------------------------------------------
-    def setup_page(self):
-        pass
