+++ spyder/plugins/switcher/widgets/item.py
